requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-ajax-selects==2.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
pycodestyle==2.7.0
PyJWT==2.1.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
